ai_code_review.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
프로젝트: GitHub PR 자동 코드 리뷰 (파일럿 구현)

이 프로젝트는 GitHub Pull Request 이벤트가 발생할 때,
1) PR에서 변경된 파일의 Diff를 확인하고,
2) 사전에 정의된 코딩 규칙과 함께
3) ChatGPT(O1) 모델을 이용하여 자동 리뷰를 수행하고,
4) 결과를 GitHub Pull Request 코멘트(Inline Comment)로 게시하는 것을 목표로 한다.

본 파일은 1차 설계 스켈레톤으로, 주요 라이브러리(PyGithub, unidiff, openai)와
GitHub Actions 환경 변수(GITHUB_TOKEN, GITHUB_REPOSITORY, PR_NUMBER, OPENAI_API_KEY)를 사용한다.

주요 흐름:
  - (1) GitHub API(PullRequest)에서 Diff 정보(파일별 patch) 획득 → unidiff로 파싱
  - (2) 코딩 규칙 로드
  - (3) ChatGPT(O1) API를 통해 코드 리뷰
  - (4) AI 응답을 분석하여 라인 단위 코멘트 데이터 생성
  - (5) GitHub PR에 In-line Review Comment로 등록

본 1차 설계의 목적:
  - 코드를 어떻게 구성/구현할지 보여주되, 실제 로직은 생략(# TODO)한다.
  - 2차 구현 단계에서 우수한 프로그래머가 이 설계를 토대로 각 함수 내용을 완성할 수 있도록 한다.
"""

import os
from typing import List, Dict, Any

# (1) PyGithub 관련
from github import Github
from github.PullRequest import PullRequest

# (2) unidiff 관련
from unidiff import PatchSet

# (3) openai 관련
import openai
from openai.types.chat.chat_completion import ChatCompletion

def main() -> None:
    """
    Main workflow, intended to run inside a GitHub Actions container.
    It relies on environment variables provided by the OS (Actions):
      - GITHUB_TOKEN (required to authenticate GitHub API calls)
      - GITHUB_REPOSITORY (e.g. "owner/repo")
      - PR_NUMBER (the pull request number to be analyzed)
      - OPENAI_API_KEY (key for ChatGPT(O1) / OpenAI API)
    """

    # 0) Load environment variables
    github_token = os.getenv("GITHUB_TOKEN")
    repo_name = os.getenv("GITHUB_REPOSITORY")  # "owner/repo"
    pr_number_str = os.getenv("PR_NUMBER")      # e.g. "123"
    openai_api_key = os.getenv("OPENAI_API_KEY")

    if not github_token or not repo_name or not pr_number_str or not openai_api_key:
        raise EnvironmentError(
            "Missing one or more required environment variables: "
            "GITHUB_TOKEN, GITHUB_REPOSITORY, PR_NUMBER, OPENAI_API_KEY."
        )

    pr_number = int(pr_number_str)

    # 1) PyGithub로 PullRequest 가져오기
    g = get_github_client(github_token)        # -> Github
    pr = get_pull_request(g, repo_name, pr_number)  # -> PullRequest

    # 2) PullRequest의 파일별 patch를 모아서 unidiff PatchSet 생성
    patch_set = get_diff_patchset(pr)          # -> PatchSet

    # 3) 코딩 규칙 로드
    rules_text = load_coding_rules()           # -> str

    # 4) ChatGPT(O1) API 호출 → 코드 리뷰 결과 획득
    review_response = get_chatgpt_review(
        patch_set=patch_set,
        rules_text=rules_text,
        openai_api_key=openai_api_key
    )                                          # -> OpenAIObject or dict

    # 5) AI 응답을 바탕으로, PR In-line 코멘트(파일 경로, 라인 번호, 내용) 생성
    comments = make_comments_from_response(
        patch_set=patch_set,
        review_response=review_response
    )                                          # -> List[Dict[str, Any]]

    # 6) GitHub PR에 코멘트 등록
    post_comments_to_pr(pr, comments)


def get_github_client(token: str) -> Github:
    """
    Create and return a PyGithub client using the provided token.

    Args:
        token (str): GitHub token for authentication

    Returns:
        Github: PyGithub client instance
    """
    # TODO: return Github(token)
    pass


def get_pull_request(g: Github, repo_name: str, pr_number: int) -> PullRequest:
    """
    Retrieve a specific PullRequest object using PyGithub.

    Args:
        g (Github): PyGithub client.
        repo_name (str): "owner/repo" string.
        pr_number (int): Pull Request number.

    Returns:
        PullRequest: The PullRequest object from PyGithub.
    """
    # TODO:
    # repo = g.get_repo(repo_name)
    # return repo.get_pull(pr_number)
    pass


def get_diff_patchset(pr: PullRequest) -> PatchSet:
    """
    From a PullRequest, gather each file's 'patch' text and parse via unidiff.

    Args:
        pr (PullRequest): The pull request object.

    Returns:
        PatchSet: Combined patch set for the entire PR.
    """
    # TODO:
    # files = pr.get_files()  # Each is PullRequestFile
    # patch_text = ""
    # for f in files:
    #    if f.patch:
    #       patch_text += f"diff --git a/{f.filename} b/{f.filename}\n"
    #       patch_text += f.patch + "\n"
    #
    # patch_set = PatchSet(patch_text)
    # return patch_set
    pass


def load_coding_rules() -> str:
    """
    Load the repository or org-level coding guidelines from a file or other source.

    Returns:
        str: The entire text of the coding rules.
    """
    # TODO: e.g. open(".github/coding-rules.md").read()
    pass


def get_chatgpt_review(
    patch_set: PatchSet,
    rules_text: str,
    openai_api_key: str
) -> ChatCompletion:
    """
    Send patch info + coding rules to ChatGPT(O1) (via openai) and return raw response.

    Args:
        patch_set (PatchSet): The unidiff PatchSet representing changed files/lines.
        rules_text (str): The loaded coding guidelines.
        openai_api_key (str): The API key for openai.

    Returns:
        ChatCompletion: The model's response object, typically from openai.ChatCompletion.create().
    """
    # TODO:
    # openai.api_key = openai_api_key
    # prompt = build_prompt_from_patchset_and_rules(patch_set, rules_text)
    # response = openai.ChatCompletion.create(...)
    # return response
    pass


def make_comments_from_response(
    patch_set: PatchSet,
    review_response: Any
) -> List[Dict[str, Any]]:
    """
    Convert the ChatGPT(O1) response into a list of comment dictionaries suitable for GitHub.

    Args:
        patch_set (PatchSet): unidiff PatchSet for reference (files/lines).
        review_response (Any): The AI's response, either an OpenAIObject or dict.

    Returns:
        List[Dict[str, Any]]: Each dict should have keys:
          {
            "path": str,
            "line": int,
            "body": str
          }
    """
    # TODO:
    # - Parse the AI's text or JSON response
    # - Possibly interpret "Line 123 in file X: ..." → map to PatchSet
    # - Return a list of comment dicts
    pass


def post_comments_to_pr(pr: PullRequest, comments: List[Dict[str, Any]]) -> None:
    """
    Post the AI-generated comments to the specified PR using PyGithub's review comment API.

    Args:
        pr (PullRequest): The PyGithub PullRequest object.
        comments (List[Dict[str, Any]]): Each dict:
            {
              "path": str,
              "line": int,
              "body": str
            }

    Returns:
        None
    """
    # TODO:
    # commit_id = pr.head.sha
    # for c in comments:
    #     pr.create_review_comment(
    #         body=c["body"],
    #         commit_id=commit_id,
    #         path=c["path"],
    #         line=c["line"]
    #     )
    pass


if __name__ == "__main__":
    main()
